+++ nextflow/modules/fastp.nf
@@ -18,16 +18,27 @@ process FASTP {
     script:
     output1 = "${module_number}-${reads[0].baseName}.fastp.fastq.gz"
     output2 = "${module_number}-${reads[1].baseName}.fastp.fastq.gz"
-    """
-    fastp -i ${reads[0]} -I ${reads[1]} \
-        -z 4 \
-        -h ${module_number}-${meta.id}_fastp.html \
-        -j ${module_number}-${meta.id}_fastp.json \
-        -R ${meta.sample}_report \
-        --failed_out ${meta.id}.fail.fastq.gz \
-        -o $output1 -O $output2
+    def check1 = file("${meta.out}/output1")
+    def check2 = file("${meta.out}/output2")
+    if (check1.exists() && check2.exists()) {
+        """
+        cp $check1.name .
+        cp $check2.name .
+        cp "${meta.log}/fastp.log" .
+        """
+    } else {
+        """
+        fastp -i ${reads[0]} -I ${reads[1]} \
+            -z 4 \
+            -h ${module_number}-${meta.id}_fastp.html \
+            -j ${module_number}-${meta.id}_fastp.json \
+            -R ${meta.sample}_report \
+            --failed_out ${meta.id}.fail.fastq.gz \
+            -o $output1 -O $output2
 
-    cp .command.out fastp.log
-    """
+        cp .command.out fastp.log
+        """
+    }
+
     //
 }
